backend/requirements.txt
fastapi==0.78.0
matplotlib==3.5.1
pandas==1.4.2
Pillow==9.2.0
pydantic==1.8.2
requests==2.27.1
starlette==0.19.1
uvicorn==0.18.2
